Adding a ci-mode

+++ hck_aux.sh
@@ -280,7 +280,7 @@ IVSHMEM_SOCKET=/tmp/ivshmem_socket_${UNIQUE_ID}
 IVSHMEM_PID=/var/run/ivshmem-server_${UNIQUE_ID}.pid
 
 run_ivshmem_server() {
-  if [ ${TEST_DEV_TYPE} = "ivshmem" ]; then
+  if [ "${TEST_DEV_TYPE}" = "ivshmem" ]; then
     echo Running ivshmem server...
     sudo rm -f /tmp/ivshmem_socket_${UNIQUE_ID}
     ${IVSHMEM_SERVER_BIN} -p ${IVSHMEM_PID}  -S ${IVSHMEM_SOCKET}
@@ -288,5 +288,8 @@ run_ivshmem_server() {
 }
 
 kill_ivshmem_server() {
-  sudo kill ${cat ${IVSHMEM_PID}}
+  if [ "${TEST_DEV_TYPE}" = "ivshmem" ]; then
+    echo stopping ivshmem server...
+    sudo kill ${cat ${IVSHMEM_PID}}
+  fi
 }
